requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
fhirclient==3.2.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
huggingface-hub==0.16.4
torch==2.0.1
transformers==4.30.2
aiohttp==3.9.1 